backend/requirements.txt
fastapi
uvicorn[standard]
beanie
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
authlib
pydantic-settings
email-validator
pymongo
httpx
python-multipart
itsdangerous